Refactor date inputs into independent component
