+++ app/hotels/src/allHotels/searchForm/SearchForm.js
@@ -1,25 +1,16 @@
 // @flow
 
 import * as React from 'react';
-import {
-  TextInput,
-  DatePicker,
-  Color,
-  Logger,
-} from '@kiwicom/react-native-app-common';
+import { TextInput, Color, Logger } from '@kiwicom/react-native-app-common';
 import { View, StyleSheet } from 'react-native';
-import moment from 'moment';
 
+import DateInput from './DateInput';
 import Guests from './guests/Guests';
 import type {
   RoomConfigurationType,
   SearchParams,
   OnChangeSearchParams,
 } from './SearchParametersType';
-
-const DISPLAY_DATE_FORMAT = 'DD/MM/YYYY';
-const MIN_NO_OF_NIGHTS = 1;
-const MAX_NO_OF_NIGHTS = 30;
 
 const styles = StyleSheet.create({
   form: {
@@ -33,13 +24,7 @@ const styles = StyleSheet.create({
     flexDirection: 'row',
     height: 40,
   },
-  datePicker: {
-    flex: 1,
-    marginRight: 10,
-  },
 });
-
-type DateInputType = 'checkin' | 'checkout';
 
 type Props = {
   location: string,
@@ -60,64 +45,12 @@ class SearchForm extends React.Component<Props> {
     this.props.onLocationChange(location);
   };
 
-  handleCheckinChange = (date: string) =>
-    this.handleDateChange(date, 'checkin');
-
-  handleCheckoutChange = (date: string) =>
-    this.handleDateChange(date, 'checkout');
-
-  handleDateChange = (datestring: string, type: DateInputType) => {
-    const date = moment(datestring, DISPLAY_DATE_FORMAT);
-    const checkin =
-      type === 'checkin' ? date : moment(this.props.search.checkin);
-    const checkout =
-      type === 'checkout' ? date : moment(this.props.search.checkout);
-    const dates = this.getValidDates(checkin, checkout, type);
-
-    if (dates) {
-      this.props.onChange(dates);
-    }
-  };
-
   handleGuestsChange = (roomsConfiguration: RoomConfigurationType) => {
     this.props.onChange({ roomsConfiguration });
   };
 
-  getValidDates = (
-    checkin: moment,
-    checkout: moment,
-    typeBeingChanged: DateInputType,
-  ) => {
-    if (checkin === null || checkout === null) {
-      return null;
-    }
-    const diff = moment(checkout).diff(checkin, 'days');
-    if (diff < MIN_NO_OF_NIGHTS || diff > MAX_NO_OF_NIGHTS) {
-      if (typeBeingChanged === 'checkout') {
-        return {
-          checkin: moment(checkout)
-            .subtract(1, 'day')
-            .toDate(),
-          checkout: checkout.toDate(),
-        };
-      } else if (typeBeingChanged === 'checkin') {
-        return {
-          checkin: checkin.toDate(),
-          checkout: moment(checkin)
-            .add(1, 'day')
-            .toDate(),
-        };
-      }
-    }
-
-    return {
-      checkin: checkin.toDate(),
-      checkout: checkout.toDate(),
-    };
-  };
-
   render = () => {
-    const { search, location } = this.props;
+    const { search, location, onChange } = this.props;
 
     return (
       <View style={styles.form}>
@@ -130,23 +63,11 @@ class SearchForm extends React.Component<Props> {
           />
         </View>
         <View style={styles.row}>
-          <DatePicker
-            placeholder="Start date"
-            format={DISPLAY_DATE_FORMAT}
-            date={search.checkin}
-            minDate={new Date()}
-            onDateChange={this.handleCheckinChange}
-            style={styles.datePicker}
-          />
-          <DatePicker
-            placeholder="End date"
-            format={DISPLAY_DATE_FORMAT}
-            date={search.checkout}
-            minDate={moment()
-              .add(1, 'day')
-              .toDate()}
-            onDateChange={this.handleCheckoutChange}
-            style={styles.datePicker}
+          <DateInput
+            containerStyles={styles.row}
+            checkin={search.checkin}
+            checkout={search.checkout}
+            onChange={onChange}
           />
           <Guests
             guests={search.roomsConfiguration}
